move files to correct locations in section
